refactor fan initialization logic and introduce FanCurveAnalyzer for RPM curve measurement
